requirements.txt
flask 
flask-cors 
elevenlabs 
python-dotenv 
google-genai